Feature: - Removed the class template nana::refer and substituted std::shared_ptr for it in the version of C++11 - Code refined.

+++ Nana.Cpp03/source/basic_types.cpp
@@ -128,10 +128,10 @@ namespace nana
 			return *this;
 		}
 
-		rectangle& rectangle::pare_off(unsigned pixels)
+		rectangle& rectangle::pare_off(int pixels)
 		{
-			x += static_cast<int>(pixels);
-			y += static_cast<int>(pixels);
+			x += pixels;
+			y += pixels;
 			width -= (pixels << 1);
 			height -= (pixels << 1);
 			return *this;
